타입 및 캐스팅 관련 경고 제거

+++ inc/color.hpp
@@ -20,12 +20,12 @@ inline glm::vec4 get_color_from_jet(float v, float vmin, float vmax) {
         c.g = 4 * (v - vmin) / dv;
     } else if (v < (vmin + 0.5 * dv)) {
         c.r = 0;
-        c.b = 1 + 4 * (vmin + 0.25 * dv - v) / dv;
+        c.b = 1 + (float)(4 * (vmin + 0.25 * dv - v) / dv);
     } else if (v < (vmin + 0.75 * dv)) {
-        c.r = 4 * (v - vmin - 0.5 * dv) / dv;
+        c.r = 4 * (float)((v - vmin - 0.5 * dv) / dv);
         c.b = 0;
     } else {
-        c.g = 1 + 4 * (vmin + 0.75 * dv - v) / dv;
+        c.g = 1 + (float)(4 * (vmin + 0.75 * dv - v) / dv);
         c.b = 0;
     }
     return glm::vec4(c, 1.0f);
